+++ CHANGELOG.md
@@ -10,9 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 - Styling issues in `CheckboxSelect` when `isValueHidden` prop is set to `true`
+- Styling some states of the button
 
 ### Refactored
 - Give `label` element in `Checkbox` and `StyledToggle` a classname to override `.styled-atlaskit-form label` in local dev
+- Icons synchronization with the Figma design
 
 ## [0.4.5]
 ### Fixed