+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.13]
+### Refactored
+- Change `div` to `p` in `addLineBreaks`
+
 ## [0.4.12]
 ### Fixed
 - `DynamicTable` fix font-size in .action-wrap